fix; rename de pice module
