+++ piece/IFile.java
@@ -1,4 +1,4 @@
-package piececomponent;
+package piece;
 
 public interface IFile {
